+++ pynumaflow/tests/function/test_sync_server.py
@@ -6,102 +6,20 @@ from grpc import StatusCode
 from grpc_testing import server_from_dictionary, strict_real_time
 from typing import Iterator, List
 
-from pynumaflow.function import (
-    Message,
-    Messages,
-    MessageT,
-    MessageTs,
-    Datum,
-    Metadata,
-)
+from pynumaflow.function import SyncServer
+from pynumaflow.tests.function.server_utils import *
 from pynumaflow.function.proto import udfunction_pb2
-from pynumaflow.function.server import UserDefinedFunctionServicer
-
-
-def map_handler(keys: List[str], datum: Datum) -> Messages:
-    val = datum.value
-    msg = "payload:%s event_time:%s watermark:%s" % (
-        val.decode("utf-8"),
-        datum.event_time,
-        datum.watermark,
-    )
-    val = bytes(msg, encoding="utf-8")
-    messages = Messages()
-    messages.append(Message(val, keys=keys))
-    return messages
-
-
-def mapt_handler(keys: List[str], datum: Datum) -> MessageTs:
-    val = datum.value
-    msg = "payload:%s event_time:%s watermark:%s" % (
-        val.decode("utf-8"),
-        datum.event_time,
-        datum.watermark,
-    )
-    val = bytes(msg, encoding="utf-8")
-    messagets = MessageTs()
-    messagets.append(MessageT(val, mock_new_event_time(), keys=keys))
-    return messagets
-
-
-async def reduce_handler(keys: List[str], datums: Iterator[Datum], md: Metadata) -> Messages:
-    interval_window = md.interval_window
-    counter = 0
-    async for _ in datums:
-        counter += 1
-    msg = (
-        f"counter:{counter} interval_window_start:{interval_window.start} "
-        f"interval_window_end:{interval_window.end}"
-    )
-    return Messages(Message(str.encode(msg), keys=keys))
-
-
-def err_map_handler(_: str, __: Datum) -> Messages:
-    raise RuntimeError("Something is fishy!")
-
-
-def err_mapt_handler(_: str, __: Datum) -> MessageTs:
-    raise RuntimeError("Something is fishy!")
-
-
-def mock_message():
-    msg = bytes("test_mock_message", encoding="utf-8")
-    return msg
-
-
-def mock_event_time():
-    t = datetime.fromtimestamp(1662998400, timezone.utc)
-    return t
-
-
-def mock_new_event_time():
-    t = datetime.fromtimestamp(1663098400, timezone.utc)
-    return t
-
-
-def mock_watermark():
-    t = datetime.fromtimestamp(1662998460, timezone.utc)
-    return t
-
-
-def mock_interval_window_start():
-    return 1662998400000
-
-
-def mock_interval_window_end():
-    return 1662998460000
-
 
 class TestServer(unittest.TestCase):
     def setUp(self) -> None:
-        my_servicer = UserDefinedFunctionServicer(
+        my_servicer = SyncServer(
             map_handler=map_handler, mapt_handler=mapt_handler, reduce_handler=reduce_handler
         )
         services = {udfunction_pb2.DESCRIPTOR.services_by_name["UserDefinedFunction"]: my_servicer}
         self.test_server = server_from_dictionary(services, strict_real_time())
 
     def test_init_with_args(self) -> None:
-        my_servicer = UserDefinedFunctionServicer(
+        my_servicer = SyncServer(
             map_handler=map_handler,
             mapt_handler=mapt_handler,
             reduce_handler=reduce_handler,
@@ -112,7 +30,7 @@ class TestServer(unittest.TestCase):
         self.assertEqual(my_servicer._max_message_size, 1024 * 1024 * 5)
 
     def test_udf_map_err(self):
-        my_servicer = UserDefinedFunctionServicer(map_handler=err_map_handler)
+        my_servicer = SyncServer(map_handler=err_map_handler)
         services = {udfunction_pb2.DESCRIPTOR.services_by_name["UserDefinedFunction"]: my_servicer}
         self.test_server = server_from_dictionary(services, strict_real_time())
 
@@ -143,7 +61,7 @@ class TestServer(unittest.TestCase):
         self.assertEqual(None, response)
 
     def test_udf_mapt_err(self):
-        my_servicer = UserDefinedFunctionServicer(mapt_handler=err_mapt_handler)
+        my_servicer = SyncServer(mapt_handler=err_mapt_handler)
         services = {udfunction_pb2.DESCRIPTOR.services_by_name["UserDefinedFunction"]: my_servicer}
         self.test_server = server_from_dictionary(services, strict_real_time())
 
@@ -277,7 +195,7 @@ class TestServer(unittest.TestCase):
 
     def test_invalid_input(self):
         with self.assertRaises(ValueError):
-            UserDefinedFunctionServicer()
+            SyncServer()
 
 
 if __name__ == "__main__":
